Rename Normalizing classes/objects
Refactor the names of members of the `normalizations` package and associated tests
* StringUtils >>> StringNormalizationUtils
* StandardEnrichmentsUtils >>> StandardNormalizations
* StringEnrichments >>> StringNormalizations
* DeDuplicationEnrichment >>> Deduplication
