+++ src/main/scala/dpla/ingestion3/enrichments/normalizations/StringNormalizations.scala
@@ -1,6 +1,6 @@
 package dpla.ingestion3.enrichments.normalizations
 
-import StringUtils._
+import StringNormalizationUtils._
 import dpla.ingestion3.model.DplaMapData.LiteralOrUri
 import dpla.ingestion3.model._
 
@@ -8,7 +8,7 @@ import dpla.ingestion3.model._
   * Universal String enrichments.
   *
   */
-class StringEnrichments {
+class StringNormalizations {
 
   /**
     * Main entry point.
