+++ src/test/scala/dpla/ingestion3/enrichments/DeduplicationTest.scala
@@ -1,13 +1,13 @@
 package dpla.ingestion3.enrichments
 
 import dpla.ingestion3.data.MappedRecordsFixture
-import dpla.ingestion3.enrichments.normalizations.DeDuplicationEnrichment
+import dpla.ingestion3.enrichments.normalizations.Deduplication
 import dpla.ingestion3.model._
 import org.scalatest.{BeforeAndAfter, FlatSpec}
 
-class DeDuplicationEnrichmentTest extends FlatSpec with BeforeAndAfter {
+class DeduplicationTest extends FlatSpec with BeforeAndAfter {
 
-  val deDuplicationEnrichment = new DeDuplicationEnrichment
+  val deDuplicationEnrichment = new Deduplication
 
   "enrich" should " remove duplicate Strings" in {
     val originalValue = Seq("foo", "foo")
